feat(cli): rename gemini binary to deepseek
