+++ package.json
@@ -53,7 +53,7 @@
     "prepublishOnly": "node scripts/prepublish.js"
   },
   "bin": {
-    "gemini": "bundle/gemini.js"
+    "deepseek": "bundle/gemini.js"
   },
   "files": [
     "bundle/",
