+++ packages/cli/package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "main": "dist/index.js",
   "bin": {
-    "gemini": "dist/index.js"
+    "deepseek": "dist/index.js"
   },
   "scripts": {
     "build": "node ../../scripts/build_package.js",
