+++ packages/cli/src/ui/components/AboutBox.tsx
@@ -32,7 +32,7 @@ export const AboutBox: React.FC<AboutBoxProps> = ({
   >
     <Box marginBottom={1}>
       <Text bold color={Colors.AccentPurple}>
-        About Gemini CLI
+        About DeepSeek CLI
       </Text>
     </Box>
     <Box flexDirection="row">
