+++ packages/cli/src/ui/components/HistoryItemDisplay.test.tsx
@@ -71,7 +71,7 @@ describe('<HistoryItemDisplay />', () => {
     const { lastFrame } = render(
       <HistoryItemDisplay {...baseItem} item={item} />,
     );
-    expect(lastFrame()).toContain('About Gemini CLI');
+    expect(lastFrame()).toContain('About DeepSeek CLI');
   });
 
   it('renders SessionSummaryDisplay for "quit" type', () => {
